==> Notebooks/02_Retrieve_twitter_data.ipynb <==
# Extracting tweets using the twitter API

* Step 1: Get `Blodgett_50_IDs.csv` file from the Datasets folder. 
* Step 2: Get access to twitter developer account and get your `twitter-api-secret.txt`.


## Importing the libraries

import tweepy
import pandas as pd

## Reading Blodgett 50k IDs and the twitter secret key

df = pd.read_csv("Blodgett_50k_IDs.csv")

#make sure to have twitter elevated access to extracted the tweets.
with open("twitter-api-secret.txt") as api_file:
  api_secret = api_file.read().splitlines()

## Authentication

auth = tweepy.AppAuthHandler(api_secret[0], api_secret[1])

api = tweepy.API(auth, wait_on_rate_limit=True)

## Dictionary where all the Tweet IDs and Tweet text will be stored

data_dict = {"Tweet_id":[],"Tweet_text":[]}

## API call to fetch the tweets and store the same in the dictionary

for tweet_id in df["Tweet_id"]:
  try:
    tweet = api.get_status(tweet_id)._json["text"]
    data_dict["Tweet_id"].append(tweet_id)
    data_dict["Tweet_text"].append(tweet)
  except: # The exception handling is done to avoid the code from crashing when a certain tweet was not found as it was deleted or removed.
    pass



## Creating data frame from the dictionary and then generating a CSV file which will be used further.

tweet_df = pd.DataFrame(data_dict)
tweet_df.to_csv("Blodgett_50k.csv")
